Fixed the frequencies and owner of the sectors in the GMMM and GMAC FIR
Fixed the sector frequencies and ownership in the GMMM and GMAC FIRs. This was done mainly to avoid confusion: when a controller is online in GMMM-ALL, it previously appeared that they were only controlling the GMMM FIR. As a result, pilots overflying GMAC believed they were outside the controller’s airspace and requested unnecessary frequency changes.